fix: chain updating issues on ssov page
move dialogs to a single component
