+++ apps/dapp/src/components/common/GlobalDialogs/index.tsx
@@ -1,8 +1,12 @@
+import ChangeNetworkDialog from '../ChangeNetworkDialog';
 import ConnectDialog from '../ConnectDialog';
+import ShareDialog from '../Share/ShareDialog';
 
 export default function GlobalDialogs() {
   return (
     <>
+      <ShareDialog />
+      <ChangeNetworkDialog />
       <ConnectDialog />
     </>
   );
