--- apps/dapp/src/components/common/Share/index.tsx
@@ -1,19 +1,0 @@
-import useShare from 'hooks/useShare';
-
-import ShareDialog from './ShareDialog';
-
-const Share = () => {
-  const open = useShare((state) => state.isOpen);
-  const handleClose = useShare((state) => state.close);
-  const shareImageProps = useShare((state) => state.shareImageProps);
-
-  return (
-    <ShareDialog
-      open={open}
-      handleClose={handleClose}
-      shareImageProps={shareImageProps}
-    />
-  );
-};
-
-export default Share;
